If we never triggered a range request but know the content length make sure to not read more data than there could be, otherwise we'd block indefinitely if the server doesn't close the connection.
related-to-bug: 2442307

+++ media/libstagefright/HTTPStream.cpp
@@ -276,11 +276,11 @@ ssize_t HTTPStream::receive(void *data, size_t size) {
             }
 
             disconnect();
-            return total == 0 ? ERROR_IO : total;
+            return total == 0 ? (ssize_t)ERROR_IO : total;
         } else if (n == 0) {
             disconnect();
 
-            return total == 0 ? ERROR_CONNECTION_LOST : total;
+            return total == 0 ? (ssize_t)ERROR_CONNECTION_LOST : total;
         }
 
         total += (size_t)n;
